feat: Implement drag tuning
Implement drag tuning, using a modified warpbox behavior. left-click drag performs a select (i.e. tune) action,
and ctrl-left-click drag performs a zoom. This behavior is the reverse of the default setting for sigplot, with
rubberbox_action set to 'select' instead of 'zoom'. Warpbox translucent style indicates a select action is underway.

Set rubberbox_mode to 'horizontal', so warpbox is constrained only in the horizontal direction while dragging, which
looks better for tuning. This behavior is also applied when drag-zooming. Therefore, if the capability to zoom on
the vertical axis is desired, rubberbox_mode sahould be set to box, either all the time or perhaps dynamiclaly only
during zoom actions.

refs #RHWEB-127

Define warpbox action to be select instead of zoom, and warpbox mode = horizontal. Now left-click selects (tunes) and ctrl-left-click zooms